split up Barcodes.swift into files for the struct/enum
